requirements.txt
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
crcmod==1.7
cryptography==46.0.3
cycler==0.12.1
Cython==3.1.4
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.14.0
fonttools==4.60.1
fsspec==2025.9.0
idna==3.11
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
ipykernel==6.29.5
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcv==2.2.0
mmcv-full==1.7.2
mmengine==0.10.7
model-index==0.1.11
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==2.2.6
omegaconf==2.3.0
opencv-python==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.3.3
parso==0.8.4
patchify==0.1.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
pycryptodome==3.23.0
pydensecrf @ file:///Users/thachha/Desktop/AIO2025-official/AIMA/OEEN/OEEM/segmentation/pydensecrf
Pygments==2.19.2
pyparsing==3.2.5
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.3
pyzmq==27.0.0
requests==2.28.2
rich==13.4.2
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
spicy==0.16.0
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2024.8.30
tomli==2.3.0
torch==2.8.0
torchvision==0.23.0
tornado==6.5.1
tqdm==4.65.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
yapf==0.43.0
zipp==3.23.0
